# api/openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Project Progress Register API
  version: 0.1.0
  description: >-
    Authoritative contract for the implemented Project Progress Register HTTP API.
    Cookie-authenticated internal API. Login returns a session-bound CSRF token;
    authenticated write requests must send it in X-CSRF-Token.
servers:
  - url: '{basePath}'
    description: Current application origin and deployment prefix.
    variables:
      basePath:
        default: /
        description: Use / for root hosting or the configured BASE_PATH, such as /backend.
...
paths:
  /api/v1/setup/bootstrap:
    post:
      tags: [Identity]
      operationId: bootstrapAdmin
      summary: Create the first Admin exactly once
      description: Available only while no users exist and BOOTSTRAP_TOKEN is configured. The token is never retained.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BootstrapRequest'}
            example: {bootstrap_token: replace-with-configured-secret, username: admin, email: admin@example.com, password: correct horse battery staple}
      responses:
        '201':
          description: Initial Admin created.
          content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}
        '403': {$ref: '#/components/responses/BootstrapDenied'}
        '404': {$ref: '#/components/responses/BootstrapUnavailable'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/auth/login:
    post:
      tags: [Identity]
      operationId: login
      summary: Start a revocable opaque session
      description: Unknown, disabled, incorrect, and throttled credentials have the same client-visible response. Sets the host-only ppr_session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
            example: {identifier: admin, password: correct horse battery staple}
      responses:
        '200':
          description: Authenticated; a new session cookie is set.
          headers:
            Set-Cookie: {schema: {type: string}, description: HttpOnly host-only ppr_session cookie with SameSite=Lax; Secure in production.}
          content: {application/json: {schema: {$ref: '#/components/schemas/SessionResponse'}}}
        '401': {$ref: '#/components/responses/InvalidCredentials'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
...
  /api/v1/admin/users:
    get:
      tags: [User administration]
      operationId: listUsers
      summary: List all internal accounts
      description: Requires an enabled Admin whose temporary password has already been replaced. Results are ordered by username.
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Complete small internal account inventory.
          content: {application/json: {schema: {$ref: '#/components/schemas/UsersResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [User administration]
      operationId: createUser
      summary: Create an Admin or Member
      description: Generates a temporary password returned only in this no-store response. The user must replace it after login.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateUserRequest'}
      responses:
        '201':
          description: Account created; copy the temporary password now because it cannot be retrieved.
          headers:
            Cache-Control: {schema: {type: string, const: no-store}}
          content: {application/json: {schema: {$ref: '#/components/schemas/CredentialResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/users/{user_id}:
    patch:
      tags: [User administration]
      operationId: updateUser
      summary: Change an account role or enabled state
      description: Uses optimistic versioning, revokes all target sessions, and refuses to remove the final enabled Admin.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/UserID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateUserRequest'}
      responses:
        '200':
          description: Account updated and target sessions revoked.
          content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/v1/admin/audit/identity:
    get:
      tags: [User administration]
      operationId: listIdentityAudit
      summary: List recent authentication and account audit events
      description: Returns the newest 200 identity, authentication, and account-authorization events. Requires an enabled Admin with no forced password change.
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Newest identity audit events in reverse chronological order.
          content: {application/json: {schema: {$ref: '#/components/schemas/IdentityAuditResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  parameters:
    CSRFToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: {type: string}
      description: Session-bound token returned by login or session recovery.
    UserID:
      in: path
      name: user_id
      required: true
      schema: {type: string, format: uuid}
...
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: ppr_session
      description: Opaque session token; only its SHA-256 digest is stored.
  responses:
    InvalidCredentials:
      description: Credentials are invalid; account existence and throttle state are intentionally undisclosed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {error: {code: invalid_credentials, message: The credentials are invalid.}}}}
    Unauthenticated:
      description: A current authenticated session is required.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    BootstrapDenied:
      description: The bootstrap secret was rejected.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    BootstrapUnavailable:
      description: Initial setup is disabled or an account already exists.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationFailed:
      description: An input field failed validation.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    UnsupportedMediaType:
      description: Content-Type does not match the operation contract (normally application/json or multipart/form-data).
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: The authenticated account lacks the required role or must replace its temporary password.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: A unique current resource already exists, an expected version is stale, or a protected invariant would be violated.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: The requested resource does not exist or is outside the caller's authorized scope.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    BootstrapRequest:
      type: object
      additionalProperties: false
      required: [bootstrap_token, username, email, password]
      properties:
        bootstrap_token: {type: string, minLength: 24, maxLength: 256, writeOnly: true}
        username: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{2,31}$'}
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 12, maxLength: 128, writeOnly: true, description: At least 12 Unicode characters and at most 128 UTF-8 bytes.}
    LoginRequest:
      type: object
      additionalProperties: false
      required: [identifier, password]
      properties:
        identifier: {type: string, maxLength: 254}
        password: {type: string, maxLength: 128, writeOnly: true}
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, email, role]
      properties:
        username: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{2,31}$'}
        email: {type: string, format: email, maxLength: 254}
        role: {type: string, enum: [admin, member]}
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [role, enabled, expected_version]
      properties:
        role: {type: string, enum: [admin, member]}
        enabled: {type: boolean}
        expected_version: {type: integer, format: int64, minimum: 1}
...
    User:
      type: object
      additionalProperties: false
      required: [id, username, email, role, enabled, must_change_password, created_at, updated_at, version]
      properties:
        id: {type: string, format: uuid}
        username: {type: string}
        email: {type: string, format: email}
        role: {type: string, enum: [admin, member]}
        enabled: {type: boolean}
        must_change_password: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        version: {type: integer, format: int64, minimum: 1}
    UserResponse:
      type: object
      additionalProperties: false
      required: [user]
      properties: {user: {$ref: '#/components/schemas/User'}}
    SessionResponse:
      type: object
      additionalProperties: false
      required: [user, csrf_token, expires_at]
      properties:
        user: {$ref: '#/components/schemas/User'}
        csrf_token: {type: string, description: Session-bound token required on authenticated writes.}
        expires_at: {type: string, format: date-time}
    UsersResponse:
      type: object
      additionalProperties: false
      required: [users]
      properties:
        users: {type: array, items: {$ref: '#/components/schemas/User'}}
    CredentialResponse:
      type: object
      additionalProperties: false
      required: [user, temporary_password]
      properties:
        user: {$ref: '#/components/schemas/User'}
        temporary_password: {type: string, readOnly: true, description: Returned once; never stored or available from later reads.}
...
    IdentityAuditRecord:
      type: object
      additionalProperties: false
      required: [id, action, target_type, outcome, occurred_at, request_id, client_ip, user_agent, details]
      properties:
        id: {type: string, format: uuid}
        actor_user_id: {type: string, format: uuid}
        actor_username: {type: string}
        action: {type: string}
        target_type: {type: string}
        target_id: {type: string, format: uuid}
        outcome: {type: string, enum: [succeeded, failed, denied]}
        occurred_at: {type: string, format: date-time}
        request_id: {type: string}
        client_ip: {type: string}
        user_agent: {type: string}
        details: {type: object, additionalProperties: true}
    IdentityAuditResponse:
      type: object
      additionalProperties: false
      required: [audit_events]
      properties:
        audit_events: {type: array, maxItems: 200, items: {$ref: '#/components/schemas/IdentityAuditRecord'}}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...